pyproject.toml
[project]
name = "license-plate-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "torch",
  "torchvision",
  "torchaudio",
  "ultralytics>=8.4.14",
  "transformers>=5.1.0",
  "fastapi>=0.128.8",
  "uvicorn>=0.40.0",
  "python-multipart>=0.0.22",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.1",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }
